+++ src/app/public/modules/type-definitions/call-signature-definition.ts
@@ -1,0 +1,15 @@
+import {
+  SkyDocsParameterDefinition
+} from './parameter-definition';
+
+import {
+  SkyDocsTypeDefinition
+} from './type-definition';
+
+export interface SkyDocsCallSignatureDefinition {
+
+  parameters: SkyDocsParameterDefinition[];
+
+  returnType: SkyDocsTypeDefinition;
+
+}
