+++ src/app/public/modules/type-definitions/class-property-definition.ts
@@ -1,0 +1,18 @@
+import {
+  SkyDocsEntryChildDefinition
+} from './entry-child-definition';
+
+/**
+ * Used to describe class properties.
+ */
+export interface SkyDocsClassPropertyDefinition extends SkyDocsEntryChildDefinition {
+
+  decorator: {
+    name: string;
+  };
+
+  defaultValue: string;
+
+  isOptional: boolean;
+
+}
