+++ src/app/public/modules/type-definitions/directive-definition.ts
@@ -1,0 +1,20 @@
+import {
+  SkyDocsClassPropertyDefinition
+} from './class-property-definition';
+
+import {
+  SkyDocsEntryDefinition
+} from './entry-definition';
+
+/**
+ * Used to describe components and directives.
+ */
+export interface SkyDocsDirectiveDefinition extends SkyDocsEntryDefinition {
+
+  eventProperties: SkyDocsClassPropertyDefinition[];
+
+  inputProperties: SkyDocsClassPropertyDefinition[];
+
+  selector: string;
+
+}
