+++ src/app/public/modules/type-definitions/entry-child-definition.ts
@@ -1,0 +1,23 @@
+import {
+  SkyDocsTypeDefinition
+} from './type-definition';
+
+/**
+ * Used to describe all entry-level children types, such as methods, properties, and accessors.
+ * @internal
+ */
+export interface SkyDocsEntryChildDefinition {
+
+  codeExample: string;
+
+  codeExampleLanguage: string;
+
+  deprecationWarning: string;
+
+  description: string;
+
+  name: string;
+
+  type: SkyDocsTypeDefinition;
+
+}
