+++ src/app/public/modules/type-definitions/entry-definition.ts
@@ -1,0 +1,19 @@
+/**
+ * Used to describe all entry-level types such as classes, directives, enumerations, etc.
+ * @internal
+ */
+export interface SkyDocsEntryDefinition {
+
+  anchorId: string;
+
+  codeExample: string;
+
+  codeExampleLanguage: string;
+
+  deprecationWarning: string;
+
+  description: string;
+
+  name: string;
+
+}
