+++ src/app/public/modules/type-definitions/enumeration-definition.ts
@@ -1,0 +1,16 @@
+import {
+  SkyDocsEntryDefinition
+} from './entry-definition';
+
+import {
+  SkyDocsEnumerationMemberDefinition
+} from './enumeration-member-definition';
+
+/**
+ * Used to describe enumerations.
+ */
+export interface SkyDocsEnumerationDefinition extends SkyDocsEntryDefinition {
+
+  members: SkyDocsEnumerationMemberDefinition[];
+
+}
