+++ src/app/public/modules/type-definitions/enumeration-member-definition.ts
@@ -1,0 +1,13 @@
+export interface SkyDocsEnumerationMemberDefinition {
+
+  codeExample: string;
+
+  codeExampleLanguage: string;
+
+  deprecationWarning: string;
+
+  description: string;
+
+  name: string;
+
+}
