+++ src/app/public/modules/type-definitions/interface-definition.component.ts
@@ -3,8 +3,14 @@ import {
   Component,
   Input
 } from '@angular/core';
-import { SkyDocsInterfaceDefinition } from './type-definitions';
-import { SkyDocsTypeDefinitionsFormatService } from './type-definitions-format.service';
+
+import {
+  SkyDocsInterfaceDefinition
+} from './interface-definition';
+
+import {
+  SkyDocsTypeDefinitionsFormatService
+} from './type-definitions-format.service';
 
 @Component({
   selector: 'sky-docs-interface-definition',
