+++ src/app/public/modules/type-definitions/interface-definition.ts
@@ -1,0 +1,16 @@
+import {
+  SkyDocsEntryDefinition
+} from './entry-definition';
+
+import {
+  SkyDocsInterfacePropertyDefinition
+} from './interface-property-definition';
+
+/**
+ * Used to describe interfaces.
+ */
+export interface SkyDocsInterfaceDefinition extends SkyDocsEntryDefinition {
+
+  properties: SkyDocsInterfacePropertyDefinition[];
+
+}
