+++ src/app/public/modules/type-definitions/interface-property-definition.ts
@@ -1,0 +1,12 @@
+import {
+  SkyDocsEntryChildDefinition
+} from './entry-child-definition';
+
+/**
+ * Used to describe interface properties.
+ */
+export interface SkyDocsInterfacePropertyDefinition extends SkyDocsEntryChildDefinition {
+
+  isOptional: boolean;
+
+}
