+++ src/app/public/modules/type-definitions/method-definition.ts
@@ -1,0 +1,16 @@
+import {
+  SkyDocsEntryChildDefinition
+} from './entry-child-definition';
+
+import {
+  SkyDocsTypeParameterDefinition
+} from './type-parameter-definition';
+
+/**
+ * Used to describe class methods.
+ */
+export interface SkyDocsClassMethodDefinition extends SkyDocsEntryChildDefinition {
+
+  typeParameters: SkyDocsTypeParameterDefinition[];
+
+}
