+++ src/app/public/modules/type-definitions/parameter-definition.ts
@@ -1,0 +1,23 @@
+import {
+  SkyDocsTypeArgumentDefinition
+} from './type-argument-definition';
+
+import {
+  SkyDocsTypeDefinition
+} from './type-definition';
+
+export interface SkyDocsParameterDefinition {
+
+  defaultValue: string;
+
+  description: string;
+
+  isOptional: boolean;
+
+  name: string;
+
+  type: SkyDocsTypeDefinition;
+
+  typeArguments: SkyDocsTypeArgumentDefinition[];
+
+}
