+++ src/app/public/modules/type-definitions/pipe-definition.ts
@@ -1,0 +1,16 @@
+import {
+  SkyDocsEntryDefinition
+} from './entry-definition';
+
+import {
+  SkyDocsClassMethodDefinition
+} from './method-definition';
+
+/**
+ * Used to describe pipes.
+ */
+export interface SkyDocsPipeDefinition extends SkyDocsEntryDefinition {
+
+  transformMethod: SkyDocsClassMethodDefinition;
+
+}
