+++ src/app/public/modules/type-definitions/type-argument-definition.ts
@@ -1,0 +1,11 @@
+import {
+  SkyDocsTypeDefinition
+} from './type-definition';
+
+export interface SkyDocsTypeArgumentDefinition {
+
+  name: string;
+
+  type: SkyDocsTypeDefinition;
+
+}
