+++ src/app/public/modules/type-definitions/type-parameter-definition.ts
@@ -1,0 +1,11 @@
+import {
+  SkyDocsTypeDefinition
+} from './type-definition';
+
+export interface SkyDocsTypeParameterDefinition {
+
+  name: string;
+
+  type: SkyDocsTypeDefinition;
+
+}
